stream tests moved to correct module
