+++ platform-sdk/swirlds-common/src/test/java/com/swirlds/common/stream/EventStreamManagerTest.java
@@ -29,7 +29,7 @@ import static org.mockito.Mockito.verify;
 import com.swirlds.common.crypto.Hash;
 import com.swirlds.common.system.NodeId;
 import com.swirlds.common.test.fixtures.RandomUtils;
-import com.swirlds.common.test.stream.ObjectForTestStream;
+import com.swirlds.common.test.fixtures.stream.ObjectForTestStream;
 import com.swirlds.test.framework.context.TestPlatformContextBuilder;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
