+++ platform-sdk/swirlds-common/src/test/java/com/swirlds/common/stream/ObjectForTestStreamTest.java
@@ -1,25 +1,27 @@
 /*
- * Copyright (C) 2021-2023 Hedera Hashgraph, LLC
+ * Copyright (C) 2023 Hedera Hashgraph, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *        http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
  */
 
-package com.swirlds.common.test.stream;
+package com.swirlds.common.stream;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import com.swirlds.common.constructable.ConstructableRegistry;
 import com.swirlds.common.test.fixtures.io.SerializationUtils;
+import com.swirlds.common.test.fixtures.stream.ObjectForTestStream;
 import java.time.Instant;
 import org.junit.jupiter.api.Test;
 
@@ -42,7 +44,7 @@ class ObjectForTestStreamTest {
 
     @Test
     void serializeDeserializeTest() throws Exception {
-        ConstructableRegistry.getInstance().registerConstructables("com.swirlds.common.test.stream");
+        ConstructableRegistry.getInstance().registerConstructables("com.swirlds.common.test.fixtures.stream");
         ObjectForTestStream deserialized = SerializationUtils.serializeDeserialize(object);
         assertEquals(object, deserialized, "deserialized object should equal to original object");
     }
