+++ platform-sdk/swirlds-common/src/test/resources/stream/readDir/2021-01-13T17_10_59.140701000Z.test_sig
