+++ platform-sdk/swirlds-common/src/testFixtures/java/com/swirlds/common/test/fixtures/stream/ObjectForTestStreamGenerator.java
@@ -1,20 +1,21 @@
 /*
- * Copyright (C) 2016-2023 Hedera Hashgraph, LLC
+ * Copyright (C) 2023 Hedera Hashgraph, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *        http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
  */
 
-package com.swirlds.common.test.stream;
+package com.swirlds.common.test.fixtures.stream;
 
 import java.time.Instant;
 import java.util.Iterator;
@@ -46,7 +47,7 @@ public class ObjectForTestStreamGenerator {
         nextTimestamp = firstTimestamp;
     }
 
-    Iterator<ObjectForTestStream> getIterator() {
+    public Iterator<ObjectForTestStream> getIterator() {
         return new Iterator<>() {
             @Override
             public boolean hasNext() {
