+++ platform-sdk/swirlds-common/src/testFixtures/java/com/swirlds/common/test/fixtures/stream/WriteToStreamConsumer.java
@@ -1,20 +1,21 @@
 /*
- * Copyright (C) 2018-2023 Hedera Hashgraph, LLC
+ * Copyright (C) 2023 Hedera Hashgraph, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *        http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
  */
 
-package com.swirlds.common.test.stream;
+package com.swirlds.common.test.fixtures.stream;
 
 import com.swirlds.common.crypto.Hash;
 import com.swirlds.common.crypto.RunningHash;
@@ -32,7 +33,7 @@ import org.apache.logging.log4j.MarkerManager;
  *
  * writes Hash and objects to a stream;
  */
-class WriteToStreamConsumer implements LinkedObjectStream<ObjectForTestStream> {
+public class WriteToStreamConsumer implements LinkedObjectStream<ObjectForTestStream> {
     /** use this for all logging, as controlled by the optional data/log4j2.xml file */
     private static final Logger logger = LogManager.getLogger(WriteToStreamConsumer.class);
 
@@ -48,9 +49,9 @@ class WriteToStreamConsumer implements LinkedObjectStream<ObjectForTestStream> {
      */
     private RunningHash runningHash;
 
-    boolean isClosed = false;
+    public boolean isClosed = false;
 
-    int consumedCount = 0;
+    public int consumedCount = 0;
 
     public WriteToStreamConsumer(SerializableDataOutputStream stream, Hash startRunningHash) throws IOException {
         this.outputStream = stream;
